Refactor TraceIndex around storage backends.
